+++ lib/taurus/qt/qtgui/taurusgui/appsettingswizard.py
@@ -45,12 +45,11 @@ import datetime
 import glob
 from lxml import etree
 
-from taurus import tauruscustomsettings
+from taurus import tauruscustomsettings, warning
 from taurus.external.qt import Qt, compat
 import taurus.qt.qtgui.panel
 import taurus.qt.qtgui.taurusgui.paneldescriptionwizard
 import taurus.qt.qtgui.input
-from taurus.core.tango.tangovalidator import TangoDeviceNameValidator
 from taurus.core.util.enumeration import Enumeration
 from taurus.qt.qtgui.util import ExternalAppAction
 
@@ -726,6 +725,12 @@ class MacroServerInfoPage(BasePage):
 
     def _getMacroServerName(self):
         if (self._macroGroupBox.isChecked()) and len(self._confWidget.macroServerComboBox.currentText()):
+            try: # TODO: tango-centric
+                from taurus.core.tango.tangovalidator import (
+                    TangoDeviceNameValidator)
+            except ImportError:
+                warning('Cannot import tango (needed for sardana integration)')
+                return None
             ms_name = str(self._confWidget.macroServerComboBox.currentText())
             complete_name, _, _ = TangoDeviceNameValidator().getNames(ms_name)
             return complete_name
@@ -735,6 +740,12 @@ class MacroServerInfoPage(BasePage):
     def _getDoorName(self):
         if (self._macroGroupBox.isChecked()) and len(self._confWidget.macroServerComboBox.currentText()):
             door_name = str(self._confWidget.doorComboBox.currentText())
+            try: # TODO: tango-centric
+                from taurus.core.tango.tangovalidator import (
+                    TangoDeviceNameValidator)
+            except ImportError:
+                warning('Cannot import tango (needed for sardana integration)')
+                return None
             complete_name, _, _ = TangoDeviceNameValidator().getNames(
                 door_name)
             return complete_name